refactor: ShazamSession 내부 정리
