+++ Segno/Segno/Data/Network/Errors/NetworkError.swift
